Move função de validação de dados cadastrais para utils
